Formations/OC-MLE/Dog-Breed-Detector/applications/tkinter_dog_breed_detector/requirements.txt
tensorflow==2.3.0
numpy==1.18.5
opencv_python==4.4.0.46
Pillow==8.1.0
PyInstaller==4.2
